Move CompareVisitor/Reporter to a public testsuite.compare module
CompareVisitor was only reachable via the internal
testsuite._internal.compare_transform module, making it awkward for
plugins to build their own pytest_register_compare_visitors hooks.
Also add a return type to the pytest_register_compare_visitors hookspec.
